Fix yank remaining line to system clipboard
In neovim, Y behaves as y$, but in vim it behaves like yy. The effect of
this is that a keymap that maps <leader>Y to "+Y will yank the entire
line to system clipboard, since keymaps like this uses vim behaviour. Is
now changed to explicitly use y$ instead